requirements.txt
aiogram==3.4.1
SQLAlchemy[asyncio]==2.0.29
aiosqlite==0.19.0
python-dotenv==1.0.1
openai>=1.35.0,<2.0.0
